rename files to snakecase
