+++ cow_py/app_data/app_data_hex.py
@@ -2,10 +2,16 @@ from typing import Dict, Any
 from web3 import Web3
 from multiformats import multibase, CID
 
-from cow_py.app_data.appDataCid import AppDataCid
-from cow_py.app_data.appDataDoc import AppDataDoc
 from cow_py.app_data.consts import DEFAULT_IPFS_READ_URI, MetaDataError
 from cow_py.app_data.utils import fetch_doc_from_cid
+
+CID_V1_PREFIX = 0x01
+CID_RAW_MULTICODEC = 0x55
+KECCAK_HASHING_ALGORITHM = 0x1B
+KECCAK_HASHING_LENGTH = 32
+CID_DAG_PB_MULTICODEC = 0x70
+SHA2_256_HASHING_ALGORITHM = 0x12
+SHA2_256_HASHING_LENGTH = 32
 
 
 class AppDataHex:
@@ -51,11 +57,11 @@ class AppDataHex:
     def _app_data_hex_to_cid(self) -> str:
         cid_bytes = self._to_cid_bytes(
             {
-                "version": 0x01,  # CIDv1
-                "multicodec": 0x55,  # Raw codec
-                "hashing_algorithm": 0x1B,  # keccak hash algorithm
-                "hashing_length": 32,  # keccak hash length (0x20 = 32)
-                "multihash_hex": self.app_data_hex,  # 32 bytes of the keccak256 hash
+                "version": CID_V1_PREFIX,
+                "multicodec": CID_RAW_MULTICODEC,
+                "hashing_algorithm": KECCAK_HASHING_ALGORITHM,
+                "hashing_length": KECCAK_HASHING_LENGTH,
+                "multihash_hex": self.app_data_hex,
             }
         )
         return multibase.encode(cid_bytes, "base16")
@@ -63,11 +69,11 @@ class AppDataHex:
     def _app_data_hex_to_cid_legacy(self) -> str:
         cid_bytes = self._to_cid_bytes(
             {
-                "version": 0x01,  # CIDv1
-                "multicodec": 0x70,  # dag-pb
-                "hashing_algorithm": 0x12,  # sha2-256 hash algorithm
-                "hashing_length": 32,  # SHA-256 length (0x20 = 32)
-                "multihash_hex": self.app_data_hex,  # 32 bytes of the sha2-256 hash
+                "version": CID_V1_PREFIX,
+                "multicodec": CID_DAG_PB_MULTICODEC,
+                "hashing_algorithm": SHA2_256_HASHING_ALGORITHM,
+                "hashing_length": SHA2_256_HASHING_LENGTH,
+                "multihash_hex": self.app_data_hex,
             }
         )
         return str(CID.decode(cid_bytes).set(version=0))
