+++ tests/app_data/app_data_hex.py
@@ -1,10 +1,8 @@
-from email.policy import HTTP
 import httpx
 import pytest
-from cow_py.app_data.appDataCid import AppDataCid
 from unittest.mock import patch
 from cow_py.app_data.consts import DEFAULT_IPFS_READ_URI, MetaDataError
-from cow_py.app_data.appDataHex import AppDataHex
+from cow_py.app_data.app_data_hex import AppDataHex
 from .mocks import (
     APP_DATA_DOC_CUSTOM,
     APP_DATA_HEX,
